+++ p2panda-net/src/sync/manager.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 
 use anyhow::{Context, Error, Result};
 use iroh_net::{Endpoint, NodeId};
-use p2panda_sync::{SyncProtocol, Topic};
+use p2panda_sync::{SyncError, SyncProtocol, Topic};
 use thiserror::Error;
 use tokio::sync::mpsc::{self, Receiver, Sender};
 use tokio::time::Duration;
@@ -61,9 +61,9 @@ enum SyncAttemptError {
     #[error("sync attempt failed due to connection or stream error")]
     Connection,
 
-    /// Error occurred while initiating a sync session.
-    #[error("sync attempt failed due to sync protocol error")]
-    Sync,
+    /// Error occurred while initiating or accepting a sync session.
+    #[error(transparent)]
+    Sync(#[from] SyncError),
 }
 
 /// An API for scheduling outbound connections and sync attempts.
@@ -237,6 +237,11 @@ where
         let sync_protocol = self.sync_protocol.clone();
         let engine_actor_tx = self.engine_actor_tx.clone();
 
+        self.active_sync_sessions
+            .entry(topic.clone())
+            .or_default()
+            .insert(peer);
+
         // Run a sync session as the initiator.
         sync::initiate_sync(
             &mut send,
@@ -246,8 +251,7 @@ where
             sync_protocol,
             engine_actor_tx,
         )
-        .await
-        .map_err(|_| SyncAttemptError::Sync)?;
+        .await?;
 
         // Clean-up the streams.
         send.finish()?;
@@ -257,24 +261,41 @@ where
         Ok(())
     }
 
-    /// Remove the given topic from the set of active sync sessions for the given peer. Reschedule
-    /// a sync attempt if the failure was caused by a connection error. Otherwise, drop the attempt
-    /// and schedule the next pending attempt.
+    /// Clean up after a failed sync attempt.
     async fn complete_failed_sync(
         &mut self,
         sync_attempt: SyncAttempt<T>,
         err: Error,
     ) -> Result<()> {
-        self.active_sync_sessions
-            .get_mut(&sync_attempt.topic)
-            .expect("active outbound sync session exists")
-            .remove(&sync_attempt.peer);
+        if let Some(session) = self.active_sync_sessions.get_mut(&sync_attempt.topic) {
+            session.remove(&sync_attempt.peer);
+        }
 
-        if let Some(SyncAttemptError::Connection) = err.downcast_ref() {
-            warn!("sync attempt failed due to connection error");
-            if sync_attempt.attempts <= MAX_RETRY_ATTEMPTS {
-                self.reschedule_attempt(sync_attempt).await?;
-                return Ok(());
+        if let Some(err) = err.downcast_ref() {
+            match err {
+                SyncAttemptError::Connection => {
+                    warn!("sync attempt failed due to connection error");
+                    if sync_attempt.attempts <= MAX_RETRY_ATTEMPTS {
+                        self.reschedule_attempt(sync_attempt).await?;
+                        return Ok(());
+                    } else {
+                        self.engine_actor_tx
+                            .send(ToEngineActor::SyncFailed {
+                                topic: Some(sync_attempt.topic),
+                                peer: sync_attempt.peer,
+                            })
+                            .await?;
+                    }
+                }
+                SyncAttemptError::Sync(err) => {
+                    warn!("sync attempt failed: {}", err);
+                    self.engine_actor_tx
+                        .send(ToEngineActor::SyncFailed {
+                            topic: Some(sync_attempt.topic),
+                            peer: sync_attempt.peer,
+                        })
+                        .await?;
+                }
             }
         }
 
@@ -286,10 +307,9 @@ where
     /// Remove the given topic from the set of active sync sessions for the given peer and add them
     /// to the set of completed sync sessions. Then schedule the next pending attempt.
     async fn complete_successful_sync(&mut self, sync_attempt: SyncAttempt<T>) -> Result<()> {
-        self.active_sync_sessions
-            .get_mut(&sync_attempt.topic)
-            .expect("active outbound sync session exists")
-            .remove(&sync_attempt.peer);
+        if let Some(session) = self.active_sync_sessions.get_mut(&sync_attempt.topic) {
+            session.remove(&sync_attempt.peer);
+        }
 
         self.completed_sync_sessions
             .entry(sync_attempt.topic)